Fix futures cancellations and generalize placement
It was not true that failed cancellation requests do not return errors
for the futures market. Sometimes they return an error code 10310. This
is now treated as success as per our rationale.

The order placement in the futures market was not general enough.
This has now been improved. I was trying to force the old data structures
into the new API. It is best to make a new tight fit for the futures API
and then explicitly translate between the two with some new functions.
This way I won't pepper the code with "loose translations".